Create lines module for estimate_lines
